This is an intermediate checkpoint (for back-up purposes) of the new transaction messaging overhaul. We're still trying to strip out VoltDB's FragmentTask and FragmentResponse messages. I also realize that we're being pretty stupid with how we serialize/deserialize ParameterSets and input data, so I'm hoping to sort out that mess here too.
git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/newdtxn-branch@2579 7a8f3d5b-8c5f-44cb-9125-2414a256df87

+++ src/frontend/org/voltdb/sysprocs/SysProcFragmentId.java
@@ -20,104 +20,104 @@ package org.voltdb.sysprocs;
 public class SysProcFragmentId
 {
     // @LastCommittedTransaction
-    public static final long PF_lastCommittedScan = 1;
-    public static final long PF_lastCommittedResults = 2;
+    public static final int PF_lastCommittedScan = 1;
+    public static final int PF_lastCommittedResults = 2;
 
     // @UpdateLogging
-    public static final long PF_updateLoggers = 3;
+    public static final int PF_updateLoggers = 3;
 
     // @Statistics
-    public static final long PF_tableData = 11;
-    public static final long PF_tableAggregator = 12;
-    public static final long PF_procedureData = 13;
-    public static final long PF_procedureAggregator = 14;
-    public static final long PF_initiatorData = 15;
-    public static final long PF_initiatorAggregator = 16;
-    public static final long PF_partitionCount = 17;
-    public static final long PF_ioData = 18;
-    public static final long PF_ioDataAggregator = 19;
+    public static final int PF_tableData = 11;
+    public static final int PF_tableAggregator = 12;
+    public static final int PF_procedureData = 13;
+    public static final int PF_procedureAggregator = 14;
+    public static final int PF_initiatorData = 15;
+    public static final int PF_initiatorAggregator = 16;
+    public static final int PF_partitionCount = 17;
+    public static final int PF_ioData = 18;
+    public static final int PF_ioDataAggregator = 19;
 
     // @Shutdown
-    public static final long PF_shutdownCommand = 21;
-    public static final long PF_procedureDone = 22;
+    public static final int PF_shutdownCommand = 21;
+    public static final int PF_procedureDone = 22;
 
     // @AdHoc
-    public static final long PF_runAdHocFragment = 31;
+    public static final int PF_runAdHocFragment = 31;
 
     // @SnapshotSave
     /*
      * Once per host confirm the file is accessible
      */
-    public static final long PF_saveTest = 40;
+    public static final int PF_saveTest = 40;
     /*
      * Agg test results
      */
-    public static final long PF_saveTestResults = 41;
+    public static final int PF_saveTestResults = 41;
     /*
      * Create and distribute tasks and targets to each EE
      */
-    public static final long PF_createSnapshotTargets = 42;
+    public static final int PF_createSnapshotTargets = 42;
     /*
      * Confirm the targets were successfully created
      */
-    public static final long PF_createSnapshotTargetsResults = 43;
+    public static final int PF_createSnapshotTargetsResults = 43;
 
     // @LoadMultipartitionTable
-    public static final long PF_loadDistribute = 50;
-    public static final long PF_loadAggregate = 51;
+    public static final int PF_loadDistribute = 50;
+    public static final int PF_loadAggregate = 51;
 
     // @SnapshotRestore
-    public static final long PF_restoreScan = 60;
-    public static final long PF_restoreScanResults = 61;
-    public static final long PF_restoreLoadReplicatedTable = 62;
-    public static final long PF_restoreLoadReplicatedTableResults = 63;
-    public static final long PF_restoreDistributeReplicatedTable = 64;
-    public static final long PF_restoreDistributePartitionedTable = 65;
-    public static final long PF_restoreDistributePartitionedTableResults = 66;
-    public static final long PF_restoreSendReplicatedTable = 67;
-    public static final long PF_restoreSendReplicatedTableResults = 68;
-    public static final long PF_restoreSendPartitionedTable = 69;
-    public static final long PF_restoreSendPartitionedTableResults = 70;
+    public static final int PF_restoreScan = 60;
+    public static final int PF_restoreScanResults = 61;
+    public static final int PF_restoreLoadReplicatedTable = 62;
+    public static final int PF_restoreLoadReplicatedTableResults = 63;
+    public static final int PF_restoreDistributeReplicatedTable = 64;
+    public static final int PF_restoreDistributePartitionedTable = 65;
+    public static final int PF_restoreDistributePartitionedTableResults = 66;
+    public static final int PF_restoreSendReplicatedTable = 67;
+    public static final int PF_restoreSendReplicatedTableResults = 68;
+    public static final int PF_restoreSendPartitionedTable = 69;
+    public static final int PF_restoreSendPartitionedTableResults = 70;
 
     // @StartSampler
-    public static final long PF_startSampler = 80;
+    public static final int PF_startSampler = 80;
 
     // @SystemInformation
-    public static final long PF_systemInformation_distribute = 90;
-    public static final long PF_systemInformation_aggregate = 91;
+    public static final int PF_systemInformation_distribute = 90;
+    public static final int PF_systemInformation_aggregate = 91;
 
     // @Quiesce
-    public static final long PF_quiesce_sites = 100;
-    public static final long PF_quiesce_processed_sites = 101;
+    public static final int PF_quiesce_sites = 100;
+    public static final int PF_quiesce_processed_sites = 101;
 
     // @SnapshotStatus
-    public static final long PF_scanSnapshotRegistries = 110;
-    public static final long PF_scanSnapshotRegistriesResults = 111;
+    public static final int PF_scanSnapshotRegistries = 110;
+    public static final int PF_scanSnapshotRegistriesResults = 111;
 
     // @SnapshotScan
-    public static final long PF_snapshotDigestScan = 124;
-    public static final long PF_snapshotDigestScanResults = 125;
-    public static final long PF_snapshotScan = 120;
-    public static final long PF_snapshotScanResults = 121;
-    public static final long PF_hostDiskFreeScan = 122;
-    public static final long PF_hostDiskFreeScanResults = 123;
+    public static final int PF_snapshotDigestScan = 124;
+    public static final int PF_snapshotDigestScanResults = 125;
+    public static final int PF_snapshotScan = 120;
+    public static final int PF_snapshotScanResults = 121;
+    public static final int PF_hostDiskFreeScan = 122;
+    public static final int PF_hostDiskFreeScanResults = 123;
 
     // @SnapshotScan
-    public static final long PF_snapshotDelete = 130;
-    public static final long PF_snapshotDeleteResults = 131;
+    public static final int PF_snapshotDelete = 130;
+    public static final int PF_snapshotDeleteResults = 131;
 
     // @InstanceId
-    public static final long PF_retrieveInstanceId = 160;
-    public static final long PF_retrieveInstanceIdAggregator = 161;
-    public static final long PF_setInstanceId = 162;
-    public static final long PF_setInstanceIdAggregator = 163;
+    public static final int PF_retrieveInstanceId = 160;
+    public static final int PF_retrieveInstanceIdAggregator = 161;
+    public static final int PF_setInstanceId = 162;
+    public static final int PF_setInstanceIdAggregator = 163;
     
     // @DatabaseDump
-    public static final long PF_dumpDistribute = 170;
-    public static final long PF_dumpAggregate = 171;
+    public static final int PF_dumpDistribute = 170;
+    public static final int PF_dumpAggregate = 171;
     
     // @RecomputeMarkovs
-    public static final long PF_recomputeMarkovsDistribute = 180;
-    public static final long PF_recomputeMarkovsAggregate = 181;
+    public static final int PF_recomputeMarkovsDistribute = 180;
+    public static final int PF_recomputeMarkovsAggregate = 181;
     
 }
